requirements.txt
requests
flask
cherrypy
sesamutils
Paste
python-dateutil